Bug fixes in Pixel

+++ src/model/Pixel.java
@@ -17,11 +17,11 @@ public class Pixel {
   }
 
   public int getIntensity() {
-    return Math.round((double) redComponent + greenComponent + blueComponent) / 3);
+    return (int) Math.round(((double) redComponent + greenComponent + blueComponent)/ 3);
   }
 
   public int getLuma() {
-    return Math.round(0.2126 * redComponent) + (0.7152 * greenComponent) + (0.0722 * blueComponent));
+    return (int) Math.round((0.2126 * redComponent) + (0.7152 * greenComponent) + (0.0722 * blueComponent));
   }
 
   public int getRedComponent() {
